[#7] Update selection method for colours
Refine algorithm to select colours from colour scheme. Add function
colourpicker to select colours from a given scheme.
Add function setbounds to determine the start and stop index in the
colour scheme. Add flag cycle to hbar, figure, and colours to cycle
through a colour scheme and start from the beginning, if more colours
are needed than in the scheme. If set to false, a BoundsError is thrown.
Update commasep to allow to set the number of digits shown.
Revise the default values in hbar and figure based on the new algorithm.

Signed-off-by: Peter Bräuer <pb866.git@gmail.com>